% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBaseline.R
\name{CheckBaseline}
\alias{CheckBaseline}
\title{Check baseline equivalency}
\usage{
CheckBaseline(raw.DF, matched.DF = NULL, treatment, variables = NULL)
}
\arguments{
\item{raw.DF}{A number.}

\item{matched.DF}{A number.}
}
\value{
Returns a list with plots and tables
}
\description{
Check baseline equivalency
}

