% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake.df}
\alias{fake.df}
\title{Fake data before and after matching}
\format{An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 5 columns.}
\usage{
fake.df
}
\description{
Fake data before and after matching
}
\keyword{datasets}

