% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBaseline.R
\name{CheckBaseline}
\alias{CheckBaseline}
\title{Check baseline equivalency}
\usage{
CheckBaseline(raw.DF, matched.DF = NULL, treatment, variables = NULL,
  names = NULL)
}
\arguments{
\item{raw.DF}{data frame you want to check.}

\item{matched.DF}{matched data frame. This one is optional.}

\item{treatment}{treatment variable}

\item{variables}{you want to check}
}
\value{
Returns a list with the baseline equivalence plot and table.
Additionally, if matched.DF is provided, it returns the propensity score distibution plot.
}
\description{
Check baseline equivalency
}
\section{Vignette}{

browseVignettes('checkbaseline')
}
\examples{
data("fake.df")
data("fake.df.matched")
X <- CheckBaseline(
  raw.DF = fake.df,
    matched.DF = fake.df.matched,
      treatment = 'Treatment',
        variables = c('V2', 'V1', 'pre.test')
        )
X$baseline.plot
}

